Lending & Borrowing

+++ contracts/Interfaces/IAurei.sol
@@ -9,13 +9,13 @@ import "../Dependencies/IERC2612.sol";
  * @notice Common interface for the Aurei token.
  */
 interface IAurei is IERC20, IERC2612 {
-
   // --- Events ---
 
-  event AureiBalanceUpdated(address _user, uint _amount);
+  event AureiBalanceUpdated(address _user, uint256 _amount);
 
   // --- Functions ---
 
   function mint(address _account, uint256 _amount) external;
+
   function burn(address _account, uint256 _amount) external;
 }
