+++ contracts/Interfaces/ILiquidation.sol
@@ -6,9 +6,7 @@ pragma solidity ^0.8.0;
  * @notice Facilitates liquidation of under-collateralized vaults.
  */
 interface ILiquidation {
-
   // --- Events ---
 
-  event Liquidation(address indexed owner, uint vaultId);
-
+  event Liquidation(address indexed owner, uint256 vaultId);
 }
